player moved to the center of minimap for better camera scaling

+++ src/init/init_camera.c
@@ -6,7 +6,7 @@
 /*   By: vmanuyko <vmanuyko@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/21 11:50:22 by vmanuyko          #+#    #+#             */
-/*   Updated: 2026/03/21 16:11:33 by vmanuyko         ###   ########.fr       */
+/*   Updated: 2026/03/24 16:38:28 by vmanuyko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,8 +17,8 @@ void	init_camera(t_camera *cam, t_user *user)
 	int	tiles;
 
 	tiles = (MINI_SIZE / MINI_TILE);
-	cam->min.x = user->player.pos.x - tiles / 2;
+	cam->min.x = (int)user->player.pos.x - tiles / 2;
 	cam->max.x = cam->min.x + tiles;
-	cam->min.y = user->player.pos.y - tiles / 2;
+	cam->min.y = (int)user->player.pos.y - tiles / 2;
 	cam->max.y = cam->min.y + tiles;
 }
